+++ ui/src/lib/get-object-changes.js
@@ -27,6 +27,7 @@ export function getObjectChanges (initialValue, updatedValue) {
           // different use case for joiningQuestions
           if (key === 'joiningQuestions') {
             changes[key] = updatedValue[key].map(q => ({ label: q.label, type: q.type }))
+              .filter(q => q.label !== '') // ensure it filters out empty entries, these will break backend
             break
           }
           // intiate the array to add, remove fields